+++ sdk/core/azure-core/test/ut/bearer_token_authentication_policy_test.cpp
@@ -16,7 +16,7 @@ private:
 public:
   explicit TestTokenCredential(
       std::shared_ptr<Azure::Core::Credentials::AccessToken const> accessToken)
-      : m_accessToken(accessToken)
+      : TokenCredential("TestTokenCredential"), m_accessToken(accessToken)
   {
   }