+++ sdk/identity/azure-identity/src/azure_cli_credential.cpp
@@ -51,9 +51,12 @@ using Azure::Identity::_detail::TokenCache;
 using Azure::Identity::_detail::TokenCredentialImpl;
 
 namespace {
-std::string const MsgPrefix = "Identity: AzureCliCredential";
+constexpr auto IdentityPrefix = "Identity: ";
+}
 
-void ThrowIfNotSafeCmdLineInput(std::string const& input, std::string const& description)
+void AzureCliCredential::ThrowIfNotSafeCmdLineInput(
+    std::string const& input,
+    std::string const& description) const
 {
   for (auto const c : input)
   {
@@ -71,20 +74,21 @@ void ThrowIfNotSafeCmdLineInput(std::string const& input, std::string const& des
         if (!std::isalnum(c, std::locale::classic()))
         {
           throw AuthenticationException(
-              MsgPrefix + ": Unsafe command line input found in " + description + ": " + input);
+              IdentityPrefix + GetCredentialName() + ": Unsafe command line input found in "
+              + description + ": " + input);
         }
     }
   }
 }
-} // namespace
-
 AzureCliCredential::AzureCliCredential(
     std::string tenantId,
     DateTime::duration cliProcessTimeout,
     Core::Credentials::TokenCredentialOptions const& options)
-    : m_tenantId(std::move(tenantId)), m_cliProcessTimeout(std::move(cliProcessTimeout))
+    : TokenCredential("AzureCliCredential"), m_tenantId(std::move(tenantId)),
+      m_cliProcessTimeout(std::move(cliProcessTimeout))
 {
   static_cast<void>(options);
+
   ThrowIfNotSafeCmdLineInput(m_tenantId, "TenantID");
 
   auto const logLevel = Logger::Level::Informational;
@@ -92,7 +96,7 @@ AzureCliCredential::AzureCliCredential(
   {
     Log::Write(
         logLevel,
-        MsgPrefix
+        IdentityPrefix + GetCredentialName()
             + " created.\n"
               "Successful creation does not guarantee further successful token retrieval.");
   }
@@ -161,7 +165,8 @@ AccessToken AzureCliCredential::GetToken(
     }
     catch (std::exception const& e)
     {
-      auto const errorMsg = MsgPrefix + " didn't get the token: \"" + e.what() + '\"';
+      auto const errorMsg
+          = IdentityPrefix + GetCredentialName() + " didn't get the token: \"" + e.what() + '\"';
 
       auto const logLevel = Logger::Level::Warning;
       if (Log::ShouldWrite(logLevel))