+++ sdk/identity/azure-identity/test/ut/azure_cli_credential_test.cpp
@@ -158,6 +158,12 @@ TEST(AzureCliCredential, Error)
   Logger::SetListener(nullptr);
 }
 
+TEST(AzureCliCredential, GetCredentialName)
+{
+  AzureCliTestCredential const cred(EmptyOutputCommand);
+  EXPECT_EQ(cred.GetCredentialName(), "AzureCliCredential");
+}
+
 TEST(AzureCliCredential, EmptyOutput)
 {
   AzureCliTestCredential const azCliCred(EmptyOutputCommand);